src/main/java/com/authorization/controller/MerchantController.java
package com.authorization.controller;

import com.authorization.model.Merchant;
import com.authorization.service.MerchantService;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/merchants")
public class MerchantController {

    private final MerchantService merchantService;

    @Autowired
    public MerchantController(MerchantService merchantService) {
        this.merchantService = merchantService;
    }

    @PostMapping
    public ResponseEntity<Merchant> createAccount(@Valid @RequestBody Merchant merchant) {
        Merchant createMerchant = merchantService.createMerchant(merchant);
        return ResponseEntity.ok(createMerchant);
    }

    @GetMapping
    public ResponseEntity<List<Merchant>> getAllMerchants() {
        List<Merchant> merchants = merchantService.getAllMerchants();
        return ResponseEntity.ok(merchants);
    }
}